requirements.txt
# ── UI ────────────────────────────────────────────────────────────
streamlit>=1.35.0
streamlit-mic-recorder>=0.0.8

# ── Speech-to-Text (local Whisper) ────────────────────────────────
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# ── Audio processing ──────────────────────────────────────────────
soundfile>=0.12.1
pydub>=0.25.1
numpy>=1.24.0

# ── LLM backends (install the ones you need) ──────────────────────

# Gemini — free tier at aistudio.google.com
google-generativeai>=0.7.0

# OpenAI — official API or OpenRouter/Together free models
openai>=1.30.0

# Groq — free tier STT + LLM at console.groq.com
groq>=0.9.0

# Ollama — local models (optional, needs Ollama installed separately)
# ollama>=0.2.0

# ── Utilities ─────────────────────────────────────────────────────
python-dotenv>=1.0.0
httpx>=0.27.0